chore(textfield): extract types
DSY-2081
